+++ test/collateral/harness.defaults.json
@@ -35,7 +35,7 @@
       "notes": "Very nice neighborhood.",
       "type": "health_center",
       "reported_date": 1550559625153,
-      "_id": "user_parent_id",
+      "_id": "parent_id_data",
       "_rev": "1-a"
     }
   },
