+++ test/harness.spec.js
@@ -1,8 +1,6 @@
-const rewire = require('rewire');
 const path = require('path');
-const sinon = require('sinon');
 const { expect } = require('chai');
-const Harness = rewire('../src/harness');
+const Harness = require('../src/harness');
 
 const formName = 'pnc_followup';
 const harness = new Harness({
@@ -227,7 +225,10 @@ describe('Harness tests', () => {
   });
 
   describe('getTasks', () => {
-    beforeEach(async () => await harness.setNow('2000-01-01'));
+    beforeEach(async () => {
+      await harness.clear();
+      await harness.setNow('2000-01-01');
+    });
 
     it('followup task present one day before schedule', async () => {
       const formResult = await harness.fillForm('pnc_followup', ['no'], ['yes', '2000-01-07']);
@@ -236,7 +237,7 @@ describe('Harness tests', () => {
       const tasks = await harness.getTasks({ now: '2000-01-07' });
       expect(tasks).to.have.property('length', 1);
       expect(tasks[0]).to.nested.include({
-        'contact._id': 'patient_id_data',
+        'contact.name': 'Patient Name',
         resolved: false,
         icon: 'newborn',
         'actions[0].form': 'pnc_followup',
@@ -283,72 +284,6 @@ describe('Harness tests', () => {
       const actual = await harness.getTasks();
       expect(actual).to.be.empty;
     });
-  });
-
-  describe('getContactSummary', () => {
-    let functionStub;
-    let basicReport;
-
-    before(async () => {
-      await harness.setNow('2000-01-01');
-      basicReport = await harness.fillForm('pnc_followup', ['no'], ['yes', '2000-01-07']);
-      expect(basicReport.errors).to.be.empty;
-    });
-    beforeEach(async () => {
-      await harness.clear();
-      basicReport.patient_id = harness.content.contact._id;
-      harness.pushMockedDoc(basicReport);
-      functionStub = sinon.stub();
-    });
-
-    it('passthrough if all args given', async () => Harness.__with__({ Function: function() { return functionStub; } })(() => {
-      const contact = {};
-      const reports = [];
-      const lineage = [];
-      harness.getContactSummary(contact, reports, lineage);
-      expect(functionStub.args[0]).to.deep.eq([contact, reports, lineage]);
-    }));
-
-
-    it('mocks datetime - setNow after start', async () => {  
-      const expectedTime = 1000;
-      await harness.setNow(expectedTime);
-      expect(new Date().getTime()).to.eq(expectedTime);
-    });
-
-    it('state used when no args given', async () => Harness.__with__({ Function: function() { return functionStub; } })(() => {
-      harness.getContactSummary();
-      const [parentOfDefault] = harness.state.contacts;
-      expect(functionStub.args[0]).to.deep.eq([harness.content.contact, harness.state.reports, [parentOfDefault]]);
-    }));
-
-    it('state used for reports and lineage but not contact', async () => Harness.__with__({ Function: function() { return functionStub; } })(() => {
-      const mockContact = { _id: 'foo' };
-      harness.getContactSummary(mockContact);
-      expect(functionStub.args[0]).to.deep.eq([mockContact, [], []]);
-    }));
-
-    it('#71 - mocked reports in state are passed to contact-summary', async () => Harness.__with__({ Function: function() { return functionStub; } })(() => {
-      const mockContact = { _id: 'foo', type: 'person' };
-      const mockReport = { _id: 'bar', patient_id: mockContact._id };
-
-      harness.pushMockedDoc(mockReport);
-      harness.getContactSummary(mockContact);
-      expect(functionStub.args[0]).to.deep.eq([mockContact, [mockReport], []]);
-    }));
-
-    it('contact summary for patient_id_data', async () => {
-      const contactSummary = harness.getContactSummary('patient_id_data');
-      
-      expect(contactSummary.cards).to.deep.eq([]);
-      expect(contactSummary.context).to.deep.eq({ muted: false, hh_contact: 'CHP Area 001 Contact' });
-      expect(contactSummary.fields).to.deep.include({ label: 'contact.age', value: '1970-07-09', filter: 'age', width: 3 });
-      expect(contactSummary.fields).to.deep.include({ label: 'Phone Number', value: '', filter: 'phone', width: 3 });
-      expect(contactSummary.fields).to.deep.include({ label: 'contact.sex', value: 'female', translate: true, width: 3 });
-      expect(contactSummary.fields).to.deep.include({ label: 'contact.external_id', value: '', width: 3 });
-    });
-
-    it('throws on invalid id', () => expect(() => harness.getContactSummary('dne')).to.throw('Cannot get summary'));
   });
 
   describe('clear', () => {
